LayoutTests/fast/events/popup-blocked-from-fake-focus.html

<html><!-- webkit-test-runner [ JavaScriptCanOpenWindowsAutomatically=false ] -->
    <head>
        <script src="../../resources/js-test.js"></script>
    </head>
    <body>
        <textarea id="test" width="100" onfocus="popup()" onblur="popup()"></textarea>
        <div id="console"></div>

        <script>
            window.jsTestIsAsync = true;
            var win;
            function popup() {
                win = window.open("about:blank", "blank");
                shouldBeNull("win");
            }


            document.getElementById("test").focus();
            document.getElementById("test").blur();

            finishJSTest();
        </script>
    </body>
</html>
